fix missing override-mode logic, start work on transposition, etc.
MIDI_StateMachine: Bug fix: I had forgotten to complete the logic of
replacing the hard-coded behavior of using the MIDI channel-volume event to
trigger "override" mode.  Fixed with this change:

68c68,69
<         if ($type == CONTROLLER() and $param == CHANNEL_VOLUME()) {
---
>         if ($type == CONTROLLER() and $param ==
>                 $self->config->filter_spec->override_cc_control_number) {

Announcer: Changed logic in END block to only unlink $tmpfile if it is
defined - avoids occasional warnings.

AppConfiguration: 'filter_spec_report': Changed the logic for retrieving
attribute names and values to allow sorting them, when constructing the
"spec report", by the numeric value - The few settings with a non-number
value are reported after the numeric settings.

ExternalCommand_MIDI_Event: Extended 'terminate' functionality to
"announce" that the program is terminating.

FilterSpecification.pm, TranspositionSpecification.pm, config1:  Began
implementation of functionality for configuration of transpositions - e.g.:

transpose_spec: 60 48..71 +7    # Middle C: transpose C3 to B4 up one fifth

The plan is for a transposition state that has been switched on to be
turned back off when the associated pitch (while in override mode) is
received.  TranspositionSpecification is a new class that stores the
parameters (bottom pitch, top pitch, # of half-steps) needed to execute a
configured transposition.

+++ Announcer.pm
@@ -45,7 +45,9 @@ sub announce {
 #####  Implementation (non-public)
 
 END {
-    unlink $tmpfile;
+    if (defined $tmpfile) {
+        unlink $tmpfile;
+    }
 }
 
 1;
